Get serialization working on Node11 (except function locations).
