+++ sdk/nodejs/tsconfig.json
@@ -58,6 +58,7 @@
         "tests/unwrap.spec.ts",
         "tests/util.ts",
         "tests/runtime/closure.spec.ts",
+        "tests/runtime/tsClosureCases.ts",
         "tests/runtime/props.spec.ts",
         "tests/runtime/langhost/run.spec.ts"
     ]
